\name{AVENGEME-package}
\alias{AVENGEME-package}
\alias{AVENGEME}
\docType{package}
\title{
\packageTitle{AVENGEME}
}
\description{
\packageDescription{AVENGEME}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{AVENGEME}
\packageIndices{AVENGEME}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{AVENGEME}

Maintainer: \packageMaintainer{AVENGEME}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
