% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avengemeShiny.R
\name{avengemeShiny}
\alias{avengemeShiny}
\title{Launch Shiny app for avengeme}
\usage{
avengemeShiny()
}
\description{
Launches a Shiny graphical user interface for the avengeme package.
}
\details{
Uses the \code{\link{estimateGeneticModel}} function to estimate all parameters that are not explicitly set in the graphical interface.
If all parameters are set, calls the \code{\link{polygenescore}} function instead.
Parameters for combined genetic and environmental risk scores are not yet implemented.
}
\references{
Palla L and Dudbridge F (2015) A fast method using polygenic scores to estimate the variance explained by genome-wide marker panels and the proportion of variants affecting a trait. Am J Hum Genet 97:250-259
}
\seealso{
\code{\link{estimatePolygenicModel}}, \code{\link{polygenescore}}
}
\author{
Frank Dudbridge
}
