% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimatePolygenicModel.R
\name{estimatePolygenicModel}
\alias{estimatePolygenicModel}
\title{Estimate Polygenic Model 
Estimate parameters of genetic model from a set of p-values for the polygenic score tested in the target sample each p-value corresponds to a polygenic score constructed from SNPs whose p-values in the training data are in a given interval possible parameters to estimate are vg1, cov12, vg2, pi01, pi02, in any combination a parameter will be estimated if its input value is unspecified or NA.}
\usage{
estimatePolygenicModel(p, nsnp, n, vg = c(NA, NA), cov12 = NA, pi0 = c(NA,
  NA), pupper = 1, nested = T, weighted = T, binary = c(F, F),
  prevalence = c(0.1, 0.1), sampling = prevalence, lambdaS = c(NA, NA),
  shrinkage = F, logrisk = F, alpha = 0.05, option = 0, boot = 0,
  bidirectional = F, initial = c(), fixvg2pi02 = F)
}
\arguments{
\item{p}{vector of p-values (or Z statistics) for polygenic scores tested in the target data automatically detects Z statistics if some entries of p are greater than 1}

\item{vg}{proportion of total variance that is explained by genetic effects in training sample by default, the variance is the same for the target sample to specify a different variance, make vg a 2-element vector}

\item{cov12}{covariance between genetic effect sizes in the two populations}

\item{pi0}{proportion of SNPs with no effects on training trait by default, the proportion is the same for the target trait to specify a different proportion, make pi0 a 2-element vector}

\item{boot}{number of bootstrap replicates to estimate approximate confidence interval if boot==0 (default), an analytic interval is calculated using profile likelihood if boot>0, a bootstrap interval is given these intervals assume that the input p-values are independent this assumption is generally untrue and the interval will be slightly smaller than it should be}

\item{bidirectional}{T if results are also given when exchanging the training and target samples in this case, vg2 can also be estimated, as can pi0 in the target sample in this case, input vector p should be twice as long with the list of p-values for training/target followed by the list for target/training}

\item{fixvg2pi02}{T if the same genetic model is assumed for the training and target samples this fixes the target variance and the covariance to both equal the variance in the training sample also fixes the proportion of null SNPs in the target sample to equal the proportion in the training sample other parameters as in polygenescore}
}
\description{
Estimate Polygenic Model 
Estimate parameters of genetic model from a set of p-values for the polygenic score tested in the target sample each p-value corresponds to a polygenic score constructed from SNPs whose p-values in the training data are in a given interval possible parameters to estimate are vg1, cov12, vg2, pi01, pi02, in any combination a parameter will be estimated if its input value is unspecified or NA.
}

