% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenescore.R
\name{polygenescore}
\alias{polygenescore}
\title{Polygenic Score:
Power, area under curve and correlation for estimated gene scores.}
\usage{
polygenescore(nsnp, n, vg1 = 0, cov12 = vg1, pi0 = 0, pupper = c(0, 1),
  nested = T, weighted = T, binary = c(F, F), prevalence = c(0.1, 0.1),
  sampling = prevalence, lambdaS = NA, shrinkage = F, logrisk = F,
  alpha = 0.05)
}
\arguments{
\item{nsnp}{number of independent SNPs in the gene score}

\item{n}{vector containing training and target sample sizes}

\item{vg1}{proportion of total variance that is explained by genetic effects in training sample}

\item{cov12}{covariance between genetic effect sizes in the two populations. This must be <=sqrt(vg1) if the genetic effects are fully correlated, set cov12=vg1 or leave unspecified, as this is the default.}

\item{pi0}{proportion of SNPs with no effects on training trait}

\item{pupper}{vector of thresholds on p-value for selection from training sample. first element is the lower bound of the first interval, second element is the upper bound of the first interval third element is the upper bound of the second interval, etc}

\item{nested}{T if the p-value intervals are nested, ie have the same lower bound if false, lower bound of the 2nd interval = upper bound of the 1st and so on. Default nested=T}

\item{weighted}{T if effect sizes used as weights in forming gene score if false, unweighted score is used, ie sum of risk alleles Default nested = T}

\item{binary}{T if training trait is binary by default, binary status of the target< trait is the same as the training trait.  If different, make binary into a vector with 2 elements}

\item{prevalence}{disease prevalence in training sample by default, prevalence is the same in the target sample If different, make prevalence a vector with 2 elements}

\item{sampling}{case/control sampling fraction in training sample by default, equals the prevalence, as in a cohort study if sampling fraction is different in the target sample, make sampling into a vector with 2 elements}

\item{lambdaS}{sibling relative recurrence risk in training sample, can be specified instead of vg1}

\item{shrinkage}{T if effect sizes are to be shrunk to BLUPs}

\item{logrisk}{T if binary trait arises from log-risk model rather than liability threshold}

\item{alpha}{type-1 error for testing association in target sample}
}
\value{
\itemize{
\item{"R2 = squared correlation between estimated gene score and target trait"}
\item{"NCP = non-centrality parameter of association test between score and target trait"}
\item{"p = expected p-value of association test"}
\item{"power = power of association test"}
\item{"AUC = for binary traits, area under ROC curve"}
\item{"MSE = for quantitative traits, mean square error between target trait and estimated gene score"}
\item{"error = error message, if any"}
}
}
\description{
Polygenic Score:
Power, area under curve and correlation for estimated gene scores.
}

