% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSizeForGeneScore.R
\name{sampleSizeForGeneScore}
\alias{sampleSizeForGeneScore}
\title{Sample Size for Genetic Score
Calculate the size of training sample for a given AUC, R2 or power targetQuantity: "AUC", "R2" or "Power"}
\usage{
sampleSizeForGeneScore(targetQuantity, targetValue, nsnp, n2 = NA, vg1 = 0,
  cov12 = vg1, pi0 = 0, weighted = T, binary = F, prevalence = 0.1,
  sampling = prevalence, lambdaS = NA, shrinkage = F, logrisk = F,
  alpha = 0.05)
}
\arguments{
\item{targetValue:}{the value of AUC/R2/Power for which to calculate sample size}

\item{n2:}{target sample size, by default set equal to the training sample size}
}
\value{
\itemize{
\item{"n = required sample size for the training sample"}
	\item{"p = p-value threshold for selecting SNPs in the score, such that the target value is achieved with the minimum sample size"}
	\item{"max = maximum AUC/R2/Power possible if the training sample contained all living humans (n1=1e10)"}
}
}
\description{
Sample Size for Genetic Score
Calculate the size of training sample for a given AUC, R2 or power targetQuantity: "AUC", "R2" or "Power"
}

